+++ lib/kumi/ir/loop/pipeline.rb
@@ -7,11 +7,18 @@ module Kumi
         module_function
 
         def default
-          @default ||= Kumi::IR::Passes::Pipeline.new([])
+          @default ||= Kumi::IR::Passes::Pipeline.new(default_passes)
         end
 
         def run(graph:, context: {})
           default.run(graph: graph, context: context)
+        end
+
+        def default_passes
+          [
+            Passes::LoopFusion.new,
+            Passes::ArrayContraction.new
+          ]
         end
       end
     end